+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# develop
+## Features:
+- [BDN-895](https://appodeal.atlassian.net/browse/BDN-895) Implemented ability to obtain banner size in BannerManager
+
 # 0.7.3
 ## Features:
 - [BDN-840](https://appodeal.atlassian.net/browse/BDN-840) Synchronize public api for Unity plugin