+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.7.9 (2025.05.**)
+## Features:
+- [BDN-931](https://appodeal.atlassian.net/browse/BDN-931) Updated Admob SDK to 24.2.0
+- [BDN-931](https://appodeal.atlassian.net/browse/BDN-931) Migrated to gradle 8
+- [BDN-931](https://appodeal.atlassian.net/browse/BDN-931) Migrated to Kotlin 2.1.0
+
+
 # 0.7.8 (2025.05.07)
 ## Features:
 - [BDN-922](https://appodeal.atlassian.net/browse/BDN-922) Updated BigoAds to 5.2.1, changed coppa setting logic